requirements.txt
absl-py==0.15.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
astunparse==1.6.3
async-generator==1.10
attrs==21.4.0
backcall==0.2.0
backports.zoneinfo==0.2.1
bleach==4.1.0
cached-property==1.5.2
cachetools==4.2.4
certifi==2022.5.18.1
cffi==1.15.0
chardet==5.0.0
charset-normalizer==3.0.1
click==8.0.4
cloudpickle==0.4.0
cmake==3.27.7
colorama==0.3.9
colored==1.4.3
colorlog==4.0.2
cycler==0.11.0
dataclasses==0.8
decorator==5.1.1
defusedxml==0.7.1
entrypoints==0.4
flatbuffers==1.12
future==0.18.2
gast==0.4.0
google-auth==2.6.6
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.34.1
h5py==2.8.0
idna==3.3
importlib-metadata==4.8.3
importlib-resources==5.4.0
iniconfig==1.1.1
ipykernel==5.5.6
ipython==7.16.3
ipython-genutils==0.2.0
ipywidgets==7.7.0
jedi==0.17.2
Jinja2==3.0.3
joblib==0.13.2
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==7.1.2
jupyter-console==6.4.3
jupyter-core==4.9.2
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.1.0
kaleido==0.2.1
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
lapjv==1.3.1
llvmlite==0.36.0
Markdown==3.3.7
MarkupSafe==2.0.1
matplotlib==3.3.4
mistune==0.8.4
mkl-fft==1.0.14
mkl-random==1.0.4
mkl-service==2.3.0
nbclient==0.5.9
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.5
notebook==6.4.10
numba==0.53.1
numexpr==2.8.1
numpy==1.19.5
oauthlib==3.2.0
opencv-python==4.2.0.34
opt-einsum==3.3.0
packaging==21.3
pandas==1.1.5
pandocfilters==1.5.0
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.4.0
plotly==5.18.0
plotpy==1.2.0
pluggy==1.0.0
prometheus-client==0.14.1
prompt-toolkit==3.0.29
protobuf==4.21.0
ptyprocess==0.7.0
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydot @ file:///tmp/build/80754af9/pydot_1615393369918/work
pydot-ng==2.0.0
Pygments==2.12.0
pyimzML==1.5.3
pyMSpec==0.1.2
pymzml==2.4.7
pyopenms==2.6.0
pyparsing==3.0.9
pyrsistent==0.18.0
pytest==7.0.1
python-dateutil @ file:///tmp/build/80754af9/python-dateutil_1626374649649/work
python-highcharts==0.4.2
PythonQwt==0.9.0
pytz==2017.2
pytz-deprecation-shim==0.1.0.post0
pyzmq==23.0.0
qtconsole==5.2.2
QtPy==2.0.1
raw2mzml==0.0.3
regex==2023.8.8
requests==2.27.1
requests-oauthlib==1.3.1
rpy2==3.1.0
rsa==4.8
scikit-learn==0.22.2.post1
scikit-multilearn==0.2.0
scipy==1.1.0
seaborn==0.11.2
Send2Trash==1.8.0
shap==0.39.0
simplegeneric==0.8.1
six==1.15.0
slicer @ file:///tmp/build/80754af9/slicer_1633422823758/work
tables==3.7.0
tenacity==8.2.2
tensorboard==2.9.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-estimator==2.5.0
tensorflow-gpu==2.5.0
termcolor==1.1.0
terminado==0.12.1
testpath==0.6.0
tomli==1.2.3
torch==1.10.2
tornado==6.1
tqdm @ file:///opt/conda/conda-bld/tqdm_1647339053476/work
traitlets==4.3.3
typing-extensions==3.7.4.3
tzdata==2024.1
tzlocal==4.2
umap-learn==0.3.10
urllib3==1.26.18
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.3
wheezy.template==3.1.0
widgetsnbextension==3.6.0
wrapt==1.12.1
zipp==3.6.0
